gateware.streambuffer: Add output double buffering
This adds a mechanism for storing frame data in a series of cyclic
buffers. We'll call them buffer A,B,C.
Initially we configure writing to B, and reading out from A.
When B is filled we shift which buffers will be used, C becomes the new
buffer for writing, and B for reading.

These addresses are latched when a read/writing sequence is started, so
there is no issue with changing these values while operations are in
progress.

If a slow input stream is used, the same frame will be output multiple
times. Alternatively if a fast input stream is used frames may simply
be skipped entirely.