firmware: Disable telemetry line
On some cameras the telemetry line may be enabled by default so we
should disable it. When enabled it outputs extra frame information
along with each frame. This confuses our image pipeline which sees an
extra line of video data, leading to data left in fifos after a
camplete frame, resulting in the image jittering on the screen

Fixes #5