gateware: Restart DMA on averager change
The Averager command on the boson has the effect of stopping the
current stream, potentially mid-frame. Then starting back up again from
the top of the new averaged frames.
Because of this cut in the stream, we need to stop our capture DMA, and
restart it.

This commit also fixes some reset logic in the DMA core itself, so now
resets actually work correctly, and will correctly terminate capturing
the current stream.

Should help with #6