scaler: Add new scaling modes, Fill/Fit
Update the DMA to enable "short" mode, this isn't ideal but it works.
Manually changing the length of DMA in firmware doesn't work great
because the change is out of sync with the scaler/framer settings.

Add new coefficients to support different scaling modes.

Correctly reset fifo0. This was likely the main cause of #6.
But I'm not claiming this issue is fixed just yet.